Move etc file generation utilities to run_common.go
